+++ NEWS.md
@@ -4,6 +4,7 @@
 - Created website
 - Ensure a single severed screen is present at any one time.
 - Check if `golem.app.prod` option is set to keep functions from running if in dev mode to allow debugging, runs in prod. If not set runs functions as expected.
+- Added `bg_image` argument to `sever` function.
 
 # sever 0.0.1